chore: rename version to v5 to make it generic
