+++ .github/workflows/ci.yaml
@@ -21,9 +21,9 @@ jobs:
         java: [17, 21]
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
       - name: Set up Java ${{ matrix.java }}
-        uses: actions/setup-java@v3
+        uses: actions/setup-java@v5
         with:
           java-version: ${{ matrix.java }}
           distribution: 'temurin'
@@ -34,9 +34,9 @@ jobs:
     name: OpenAPI Linter
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
+        uses: actions/checkout@v5
+      - uses: actions/setup-node@v5
         with:
           node-version: 18
-      - name: v4 Lint
-        run: npx @redocly/cli lint api/v4/openapi.yaml
+      - name: v5 Lint
+        run: npx @redocly/cli lint api/v5/openapi.yaml
