+++ .github/workflows/cicd.yaml
@@ -15,12 +15,12 @@ jobs:
     name: OpenAPI Linter
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
+        uses: actions/checkout@v5
+      - uses: actions/setup-node@v5
         with:
           node-version: 18
-      - name: v4 Lint
-        run: npx @redocly/cli lint api/v4/openapi.yaml
+      - name: v5 Lint
+        run: npx @redocly/cli lint api/v5/openapi.yaml
 
   deploy-latest-packages:
     if: github.repository_owner == 'guacsec' && !contains(github.event.head_commit.message, 'build(release)')
@@ -31,9 +31,9 @@ jobs:
     name: Deploy
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
       - name: Set up Java 17
-        uses: actions/setup-java@v3
+        uses: actions/setup-java@v5
         with:
           java-version: '17'
           distribution: 'temurin'
