+++ src/filtered/js-api/package.json
@@ -1,16 +1,16 @@
 {
   "name": "@guacsec/${project.artifactId}",
   "version": "${project.version}",
-  "description": "Trustify Vulnerability Analysis :: Exhort :: Javascript API",
+  "description": "Trustify :: Dependency Analytics :: API",
   "type": "commonjs",
   "files": [
     "model/"
   ],
   "exports": {
-    "./model/v4/*": {
-      "require": "./model/v4/*",
-      "import": "./model/v4/*",
-      "types": "./model/v4/*"
+    "./model/v5/*": {
+      "require": "./model/v5/*",
+      "import": "./model/v5/*",
+      "types": "./model/v5/*"
     }
   },
   "publishConfig": {
@@ -28,7 +28,8 @@
   "keywords": [
     "analysis",
     "codeready",
-    "exhort",
+    "trustify",
+    "dependency-analytics",
     "secure",
     "supply-chain",
     "vulnerability"
