+++ src/main/java/com/github/guacsec/trustifyda/api/PackageRef.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2025 Trustify Exhort Authors
+ * Copyright 2023-2025 Trustify Dependency Analytics Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package com.github.guacsec.exhort.api;
+package com.github.guacsec.trustifyda.api;
 
 import java.util.Objects;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.github.guacsec.exhort.api.serialization.PackageURLSerializer;
+import com.github.guacsec.trustifyda.api.serialization.PackageURLSerializer;
 import com.github.packageurl.MalformedPackageURLException;
 import com.github.packageurl.PackageURL;
