+++ src/main/java/com/github/guacsec/trustifyda/api/v5/SeverityUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2025 Trustify Exhort Authors
+ * Copyright 2023-2025 Trustify Dependency Analytics Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.github.guacsec.exhort.api.v4;
+package com.github.guacsec.trustifyda.api.v5;
 
 public class SeverityUtils {
